Function index

This page lists the docstrings for each public function. You can also see this information in the Julia REPL by typing ? then the name of the function.

Convenience functions

These functions make use of the basic functions to allow you to calculate specific quantities easily.

Geodesics.angular_distance — Function

angular_distance(lon0, lat0, lon1, lat1, degrees=true; f=F_WGS84) -> Δ

Return the angular distance between points (lon0,lat0) and (lon1,lat1) on a flattened sphere. The default flattening is 0.0033528106718309896. By default, input and output are in degrees, but specify degrees as false to use radians. Use f=0 to perform computations on a sphere.

Geodesics.surface_distance — Function

surface_distance(lon0, lat0, lon1, lat1, a, degrees::Bool=false; f=F_WGS84)

Return the physical distance between points (lon0,lat0) and (lon1,lat1) on the flattened sphere with flattening f and semimajor radius a. Distance is given in the same units as a. By default, input angles are in degrees, but specify degrees as false to use radians. Use f=0 to perform computations on a sphere.

Geodesics.angular_step — Function

step(lon, lat, azimuth, distance, degrees=false; f=F_WGS84) -> lon′, lat′, backazimuth

Return the longitude lon′, latitude lat′ and backazimuth baz reached by travelling an angular distance along azimuth from the starting point at (lon,lat). The default flattening is 0.0033528106718309896. By default, input and output are in degrees, but specify degrees as false to use radians. Use f=0 to perform computations on a sphere.

Geodesics.azimuth — Function

azimuth(lon0, lat0, lon1, lat1, degrees=true, f=F_WGS84)

Return the azimuth from point (lon0,lat0) to point (lon1,lat1) on a flattened sphere. The default flattening is 0.0033528106718309896. By default, input and output are in degrees, but specify degrees as false to use radians. Use f=0 to perform computations on a sphere.

Basic functions

These functions solve the geodesic inverse or forward problems, returning all available information.

Geodesics.forward — Function

forward(lon, lat, azimuth, distance, a, f) -> lon′, lat′, backazimuth

Return the longitude lon′ and latitude lat′ and backazimuth of a projected point, reached by travelling along an azimuth for distance from an original point at (lon, lat). Specify the spheroid with the semimajor radius a and flattening f.

Coordinates and azimuth are in radians.

Calculations use Vincenty's forward formula [1].

References

Vincenty, T. (1975). "Direct and Inverse Solutions of Geodesics on the Ellipsoid with application of nested equations" (PDF). Survey Review. XXIII (176): 88–93. doi:10.1179/sre.1975.23.176.88

Geodesics.inverse — Function

inverse(lon1, lat1, lon2, lat2, a, f) -> distance, azimuth, backazimuth

Return the distance, azimuth and backazimuth between two points with longitudes lon1 and lon2, and latitudes lat1 and lat2. Specify the spheroid with the semimajor radius a and flattening f.

Coordinates and angles are in radians, whilst distance is in the same units as a.

Calculations use Vincenty's inverse formula [1].

References

Vincenty, T. (1975). "Direct and Inverse Solutions of Geodesics on the Ellipsoid with application of nested equations" (PDF). Survey Review. XXIII (176): 88–93. doi:10.1179/sre.1975.23.176.88